move RegionalRates to market, belongs there
