+++ src/server.rs
@@ -1,4 +1,5 @@
 use crate::aws::Aws;
+use crate::market::RegionalRates;
 use anyhow::Result;
 use serde::{Deserialize, Serialize};
 use std::fs;
@@ -9,18 +10,6 @@ use std::net::{TcpListener, TcpStream};
 struct Spec {
     allowed_regions: Vec<String>,
     min_rates: Vec<RegionalRates>,
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
-pub struct RateCard {
-    pub instance: String,
-    pub min_rate: i64,
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
-pub struct RegionalRates {
-    pub region: String,
-    pub rate_cards: Vec<RateCard>,
 }
 
 async fn handle_read(
